requirements.txt
torch
torchvision
gradio
Pillow
numpy
opencv-python-headless
